fix: correct data types for progress tracking fields in BackupJob model
Fixed type mismatch between database schema and SQLAlchemy model that
was preventing progress_percent values from being stored correctly.

Root cause:
- Database migration created progress_percent as REAL (float)
- Model defined it as Integer, causing type conversion issues
- File count and current_file worked because they matched types
- Progress percentage always showed 0.0 despite backend updates

Changes to app/database/models.py:
- Import Float and BigInteger from sqlalchemy
- Change progress_percent: Column(Integer) → Column(Float, default=0.0)
- Change size fields to BigInteger to match database BIGINT type:
  - original_size: Integer → BigInteger
  - compressed_size: Integer → BigInteger
  - deduplicated_size: Integer → BigInteger

This ensures SQLAlchemy correctly reads/writes float values for progress
tracking, enabling real-time progress updates during backup operations.

+++ app/database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, Text, ForeignKey
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, Text, ForeignKey, Float, BigInteger
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.database.database import Base
@@ -114,11 +114,11 @@ class BackupJob(Base):
     log_file_path = Column(String, nullable=True)  # Path to streaming log file
 
     # Detailed progress fields from Borg JSON output
-    original_size = Column(Integer, default=0)  # Original uncompressed size in bytes
-    compressed_size = Column(Integer, default=0)  # Compressed size in bytes
-    deduplicated_size = Column(Integer, default=0)  # Deduplicated size in bytes
+    original_size = Column(BigInteger, default=0)  # Original uncompressed size in bytes
+    compressed_size = Column(BigInteger, default=0)  # Compressed size in bytes
+    deduplicated_size = Column(BigInteger, default=0)  # Deduplicated size in bytes
     nfiles = Column(Integer, default=0)  # Number of files processed
     current_file = Column(Text, nullable=True)  # Current file being processed
-    progress_percent = Column(Integer, default=0)  # Progress percentage
+    progress_percent = Column(Float, default=0.0)  # Progress percentage
 
     created_at = Column(DateTime, default=datetime.utcnow)
